fix bug preventing unauth'd users from adding hack
when an unauthenticated user submits a hack, we should redirect them to
the github login page, and then after they've logged in, we should add
the hack they tried to submit.

however, because the ajax client was not authenticated at the beginning
of this process, the post-login ajax call fails with an HTTP 401.

to fix this, we don't hold on to a reference to the ajax client, we just
get a new reference when making the post-login ajax call.