use ellipsis for long titles
now that we have some actual user data, people are putting in longer
hacks than I expected, which messes up the UI a bit.

to fix this, when the project is not selected, we show an ellipsis (...)
if the text would overflow the allotted space. when the project is
selected, we show all the text, and just let it overflow.

in addition, I've added a maxlength attribute (40 characters) to the
input to prevent longer titles. I think 40 should be enough for hack
titles, but we can adjust it if it is too short.