use hash-based routing for historical event urls
we can't use html5 based routing, because then on refresh (or initial
navigation), github would attempt to serve the HTML5 url, and respond
with a 404. It appears that node.js servers solve this by routing all
urls to the root page, so then the app would be served, and it could
examine the url.

However, we are hosted on github pages, so we can't apply that
configuration. So we'll use hash-based routing.